extract keymaps.lua file
